refactor(userspace): extract policy/filter wire types to protocol_policies.go
Sixth code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the security-policy, firewall-filter, policer and
application-catalog configuration snapshots into a cohesive sibling
protocol_policies.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change):
FirewallFilterSnapshot, FirewallTermSnapshot, FlexMatchSnapshot,
PolicerSnapshot, ThreeColorPolicerSnapshot, PolicyApplicationSnapshot,
AppCatalogEntrySnapshot, PolicyRuleSnapshot.

Behavior byte-identical; json tags preserved for the Rust userspace-dp
deserializer. Build passes; gofmt clean.
